create inheritance from BallpointPen to Pen, to allow configuration of alternative pens.
